+++ drivers/vga-nextgen/vga.c
@@ -311,15 +311,15 @@ void __time_critical_func() dma_handler_VGA() {
         case GRAPHICSMODE_DEFAULT:
             input_buffer_8bit = input_buffer + y * width;
             for (int i = width; i--;) {
-                uint8_t t = *input_buffer_8bit++ >> 6;
+                uint8_t t = *input_buffer_8bit++ >> 4;
                 *output_buffer_16bit++ = current_palette[t];
                 *output_buffer_16bit++ = current_palette[t];
             }
             break;
         case GRAPHICSMODE_ASPECT:
             input_buffer_8bit = input_buffer + y * width;
-            for (int i = width; i--;) {
-                *output_buffer_16bit++ = current_palette[(*input_buffer_8bit++) >> 6];
+            for (int x = 0; x< width; x++) {
+                *output_buffer_16bit++ = current_palette[*input_buffer_8bit++ >> 4];
             }
             break;
         default: