+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.1.0
+
+- Save/Load
+- TV-Link supervision background
+- Palette names
+
 # v1.0.0
 
 - Everything is working